backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
psycopg2-binary
stripe
plaid-python
boto3
werkzeug
gunicorn
